+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ### v0.11.0 (Date TBD)
 
-Fixed issues: https://github.com/eclipse/lsp4j/milestone/18?closed=1
+ * Added new module `org.eclipse.lsp4j.websocket.jakarta` for using LSP4J over Jakarta websockets
 
+Fixed issues: https://github.com/eclipse/lsp4j/milestone/18?closed=1
 
 ### v0.10.0 (Nov. 2020)